Move fixer implementations to linter module
