+++ silence_lint_error/linters/ruff.py
@@ -57,3 +57,22 @@ class Ruff:
         [rule_name] = {violation.rule_name for violation in violations}
         linenos_to_silence = {violation.lineno for violation in violations}
         return comments.add_noqa_comments(src, linenos_to_silence, rule_name)
+
+    def remove_silence_comments(self, src: str, rule_name: RuleName) -> str:
+        return comments.remove_error_silencing_comments(
+            src, comment_type='noqa', error_code=rule_name,
+        )
+
+    def apply_fixes(
+            self, rule_name: RuleName, filenames: Sequence[str],
+    ) -> tuple[int, str]:
+        proc = subprocess.run(
+            (
+                sys.executable, '-mruff',
+                'check', '--fix',
+                '--select', rule_name,
+                *filenames,
+            ),
+            capture_output=True, text=True,
+        )
+        return proc.returncode, proc.stdout.strip()
